restructure files and folders of pools module
