+++ src/modules/pools/bond/index.jsx
@@ -4,7 +4,7 @@ import DateLib from "@/lib/date/DateLib";
 import { Label } from "@/components/UI/atoms/label";
 import { RegularButton } from "@/components/UI/atoms/button/regular";
 import { Container } from "@/components/UI/atoms/container";
-import { BondInfoCard } from "@/components/UI/organisms/pools/bond/BondInfoCard";
+import { BondInfoCard } from "@/src/modules/pools/bond/BondInfoCard";
 import { mergeAlternatively } from "@/utils/arrays";
 import { TokenAmountInput } from "@/components/UI/organisms/token-amount-input";
 import { ReceiveAmountInput } from "@/components/UI/organisms/receive-amount-input";
