+++ src/modules/pools/staking/StakeForm.jsx
