+++ src/pages/pools/bond/index.jsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
-import { PoolsTabs } from "@/src/modules/pools/bond/PoolsTabs";
-import BondPage from "@/components/pages/pools/bond";
+import { PoolsTabs } from "@/src/modules/pools/PoolsTabs";
+import BondPage from "@/src/modules/pools/bond";
 import { ComingSoon } from "@/components/pages/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
