+++ src/pages/pools/pod-staking.jsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 
-import { PoolsTabs } from "@/src/modules/pools/bond/PoolsTabs";
-import { PodStakingPage } from "@/components/pages/pools/pod-staking";
+import { PoolsTabs } from "@/src/modules/pools/PoolsTabs";
+import { PodStakingPage } from "@/src/modules/pools/pod-staking";
 import { ComingSoon } from "@/components/pages/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
